third_party/blink/web_tests/platform/generic/jquery/offset-expected.txt

ALERT: 387 tests of 387 passed, 0 failed.
